+++ test/output_csharp/src/phmap_parallel_flat_hash_map_int_float.cs
@@ -118,10 +118,10 @@ public static partial class MR
                 public Const_ParallelFlatHashMap_Int_Float(ParallelFlatHashMap_Int_Float other) : this((Const_ParallelFlatHashMap_Int_Float)other) {}
 
                 /// The number of elements.
-                public unsafe ulong Size()
+                public unsafe nuint Size()
                 {
                     [System.Runtime.InteropServices.DllImport("bleh", EntryPoint = "MR_C_phmap_parallel_flat_hash_map_int_float_size", ExactSpelling = true)]
-                    extern static ulong __MR_C_phmap_parallel_flat_hash_map_int_float_size(_Underlying *_this);
+                    extern static nuint __MR_C_phmap_parallel_flat_hash_map_int_float_size(_Underlying *_this);
                     MR.CS.Misc._Exceptions.Prepare();
                     var __c_ret = __MR_C_phmap_parallel_flat_hash_map_int_float_size(_UnderlyingPtr);
                     MR.CS.Misc._Exceptions.ThrowIfNeeded();
