+++ test/output_csharp/src/std_multiset_int.cs
@@ -118,10 +118,10 @@ public static partial class MR
                 public Const_Multiset_Int(Multiset_Int other) : this((Const_Multiset_Int)other) {}
 
                 /// Construct from a range of elements.
-                public unsafe Const_Multiset_Int(int? ptr, ulong size) : this(null, is_owning: true)
+                public unsafe Const_Multiset_Int(int? ptr, nuint size) : this(null, is_owning: true)
                 {
                     [System.Runtime.InteropServices.DllImport("bleh", EntryPoint = "MR_C_std_multiset_int_ConstructFromRange", ExactSpelling = true)]
-                    extern static MR.CS.Std.Multiset_Int._Underlying *__MR_C_std_multiset_int_ConstructFromRange(int *ptr, ulong size);
+                    extern static MR.CS.Std.Multiset_Int._Underlying *__MR_C_std_multiset_int_ConstructFromRange(int *ptr, nuint size);
                     MR.CS.Misc._Exceptions.Prepare();
                     int __deref_ptr = ptr.GetValueOrDefault();
                     _UnderlyingPtr = __MR_C_std_multiset_int_ConstructFromRange(ptr.HasValue ? &__deref_ptr : null, size);
@@ -129,10 +129,10 @@ public static partial class MR
                 }
 
                 /// The number of elements.
-                public unsafe ulong Size()
+                public unsafe nuint Size()
                 {
                     [System.Runtime.InteropServices.DllImport("bleh", EntryPoint = "MR_C_std_multiset_int_size", ExactSpelling = true)]
-                    extern static ulong __MR_C_std_multiset_int_size(_Underlying *_this);
+                    extern static nuint __MR_C_std_multiset_int_size(_Underlying *_this);
                     MR.CS.Misc._Exceptions.Prepare();
                     var __c_ret = __MR_C_std_multiset_int_size(_UnderlyingPtr);
                     MR.CS.Misc._Exceptions.ThrowIfNeeded();
@@ -151,10 +151,10 @@ public static partial class MR
                 }
 
                 /// Checks if the contain contains this key.
-                public unsafe ulong Count(int key)
+                public unsafe nuint Count(int key)
                 {
                     [System.Runtime.InteropServices.DllImport("bleh", EntryPoint = "MR_C_std_multiset_int_count", ExactSpelling = true)]
-                    extern static ulong __MR_C_std_multiset_int_count(_Underlying *_this, int *key);
+                    extern static nuint __MR_C_std_multiset_int_count(_Underlying *_this, int *key);
                     MR.CS.Misc._Exceptions.Prepare();
                     var __c_ret = __MR_C_std_multiset_int_count(_UnderlyingPtr, &key);
                     MR.CS.Misc._Exceptions.ThrowIfNeeded();
@@ -481,10 +481,10 @@ public static partial class MR
                 }
 
                 /// Construct from a range of elements.
-                public unsafe Multiset_Int(int? ptr, ulong size) : this(null, is_owning: true)
+                public unsafe Multiset_Int(int? ptr, nuint size) : this(null, is_owning: true)
                 {
                     [System.Runtime.InteropServices.DllImport("bleh", EntryPoint = "MR_C_std_multiset_int_ConstructFromRange", ExactSpelling = true)]
-                    extern static MR.CS.Std.Multiset_Int._Underlying *__MR_C_std_multiset_int_ConstructFromRange(int *ptr, ulong size);
+                    extern static MR.CS.Std.Multiset_Int._Underlying *__MR_C_std_multiset_int_ConstructFromRange(int *ptr, nuint size);
                     MR.CS.Misc._Exceptions.Prepare();
                     int __deref_ptr = ptr.GetValueOrDefault();
                     _UnderlyingPtr = __MR_C_std_multiset_int_ConstructFromRange(ptr.HasValue ? &__deref_ptr : null, size);
@@ -492,10 +492,10 @@ public static partial class MR
                 }
 
                 /// Assign from a range of elements, overwriting previous contents.
-                public unsafe void Assign(int? ptr, ulong size)
+                public unsafe void Assign(int? ptr, nuint size)
                 {
                     [System.Runtime.InteropServices.DllImport("bleh", EntryPoint = "MR_C_std_multiset_int_AssignFromRange", ExactSpelling = true)]
-                    extern static void __MR_C_std_multiset_int_AssignFromRange(_Underlying *_this, int *ptr, ulong size);
+                    extern static void __MR_C_std_multiset_int_AssignFromRange(_Underlying *_this, int *ptr, nuint size);
                     MR.CS.Misc._Exceptions.Prepare();
                     int __deref_ptr = ptr.GetValueOrDefault();
                     MR.CS.Misc._Exceptions.ThrowIfNeeded();
