+++ test/output_csharp/src/std_string.cs
@@ -71,10 +71,10 @@ public static partial class MR
                 public static unsafe implicit operator Const_String(string other) {return new(other);}
 
                 /// The number of characters in the string, excluding the null-terminator.
-                public unsafe ulong Size()
+                public unsafe nuint Size()
                 {
                     [System.Runtime.InteropServices.DllImport("bleh", EntryPoint = "MR_C_std_string_size", ExactSpelling = true)]
-                    extern static ulong __MR_C_std_string_size(_Underlying *_this);
+                    extern static nuint __MR_C_std_string_size(_Underlying *_this);
                     MR.CS.Misc._Exceptions.Prepare();
                     var __c_ret = __MR_C_std_string_size(_UnderlyingPtr);
                     MR.CS.Misc._Exceptions.ThrowIfNeeded();
